seperate camera logic to simple camera controller
