+++ applications/sandbox/module/testModule.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <core/core.hpp>
 #include "../systems/testsystem.hpp"
+#include "../systems/simplecameracontroller.hpp"
 
 #include <core/math/math.hpp>
 
@@ -14,8 +15,8 @@ public:
         addProcessChain("TestChain");
 
         reportComponentType<sah>();
-        reportComponentType<addRB>();
         reportSystem<TestSystem>();
+        reportSystem<SimpleCameraController>();
     }
 
     virtual priority_type priority() override
