+++ applications/sandbox/systems/testsystem.hpp
@@ -47,14 +47,6 @@ enum class PhysicsUnitTestMode
     CollisionResolution
 };
 
-struct player_move : public app::input_axis<player_move> {};
-struct player_strive : public app::input_axis<player_strive> {};
-struct player_fly : public app::input_axis<player_fly> {};
-struct player_look_x : public app::input_axis<player_look_x> {};
-struct player_look_y : public app::input_axis<player_look_y> {};
-
-struct player_hover : public app::input_axis< player_hover> {};
-
 // Move and strive for the wireframe sphere (which holds the only audio source)
 // For testing the movement of audio sources (Spatial audio/doppler)
 struct audio_move : public app::input_axis<audio_move> {};
@@ -68,25 +60,14 @@ struct stop_audio_source : public app::input_action<stop_audio_source> {};
 struct rewind_audio_source : public app::input_action<rewind_audio_source> {};
 struct audio_test_input : public app::input_action<audio_test_input> {};
 
-struct exit_action : public app::input_action<exit_action> {};
-
-struct fullscreen_action : public app::input_action<fullscreen_action> {};
-struct escape_cursor_action : public app::input_action<escape_cursor_action> {};
-struct vsync_action : public app::input_action<vsync_action> {};
-
 //some test stuff just so i can change things
 
 struct physics_test_move : public app::input_axis<physics_test_move> {};
 
-
-struct activate_CRtest0 : public app::input_action<activate_CRtest0> {};
-struct activate_CRtest1 : public app::input_action<activate_CRtest1> {};
 struct activate_CRtest2 : public app::input_action<activate_CRtest2> {};
 struct activate_CRtest3 : public app::input_action<activate_CRtest3> {};
 
 struct activateFrictionTest : public app::input_action<activateFrictionTest> {};
-
-
 
 struct extendedPhysicsContinue : public app::input_action<extendedPhysicsContinue> {};
 struct nextPhysicsTimeStepContinue : public app::input_action<nextPhysicsTimeStepContinue> {};
@@ -100,7 +81,6 @@ public:
         app::WindowSystem::requestWindow(world_entity_id, math::ivec2(1360, 768), "LEGION Engine", "Legion Icon", nullptr, nullptr, 1); // Create the request for the main window.
     }
 
-    ecs::entity_handle player;
     ecs::entity_handle audioSphereLeft;
     ecs::entity_handle audioSphereRight;
 
@@ -125,165 +105,10 @@ public:
 
     virtual void setup()
     {
-//#pragma region OpenCL
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//        std::vector<int> ints;
-//
-//
-//        // load a file with tons of numbers
-//        auto res = fs::view("assets://bigint.txt").get();
-//        if (res == common::valid) {
-//
-//            filesystem::basic_resource contents = res;
-//
-//            char* buf = new char[6]{0};
-//
-//            // get a lot of integers from the data
-//            for (size_t i = 0; i < contents.size() && i < 5 * 2048; i += 5)
-//            {
-//                memcpy(buf, contents.data() + i, 5);
-//                ints.push_back(std::strtol(buf,nullptr,10));
-//            }
-//
-//            delete[] buf;
-//        }
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//        // split the integers into two datasets
-//        std::vector<int> first_ints(ints.begin(), ints.begin() + ints.size() / 2);
-//        std::vector<int> second_ints(ints.begin() + ints.size() / 2, ints.end());
-//
-//        // check how many things we need to process
-//        size_t to_process = std::min(first_ints.size(), second_ints.size());
-//
-//        // get an appropriately sized output buffer
-//        std::vector<int> results(to_process);
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//        // import some names
-//        using compute::in,compute::out;
-//
-//        log::debug("Loading OpenCL kernel");
-//
-//        // create opencl-"function"
-//        auto vector_add = fs::view("assets://kernels/vadd_kernel.cl").load_as<compute::function>("vector_add");
-//
-//        log::debug ("Done loading OpenCL kernel");
-//
-//        #if 0
-//
-//        // invoke "function"
-//        auto code1 = vector_add(to_process,first_ints,second_ints,out(results));
-//
-//        // output results
-//        if (code1.valid())
-//            for (int& i : results)
-//            {
-//                log::info("got {}", i);
-//            }
-//        #endif
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//        log::debug("Creating hardware buffers");
-//
-//        // If we need to reuse the function it is more efficient to keep the buffers around.
-//        // This way we do not allocate resources on our computation device every-time.
-//        auto A = compute::Context::createBuffer(first_ints,     compute::buffer_type::READ_BUFFER,  "A");
-//        auto B = compute::Context::createBuffer(second_ints,    compute::buffer_type::READ_BUFFER,  "B");
-//        auto C = compute::Context::createBuffer(results,        compute::buffer_type::WRITE_BUFFER, "C");
-//
-//        log::debug("Done creating hardware buffers");
-//
-//        log::debug("Executing kernel");
-//        // the invocation is otherwise exactly the same
-//        auto code2 = vector_add(to_process,A,B,C);
-//
-//        log::debug("Done executing kernel");
-//
-//        // output results
-//        if (code2.valid())
-//            for (int& i : results)
-//            {
-//                log::info("got {}", i);
-//            }
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//#pragma endregion
 
 #pragma region Input binding
         app::InputSystem::createBinding<physics_test_move>(app::inputmap::method::LEFT, -1.f);
         app::InputSystem::createBinding<physics_test_move>(app::inputmap::method::RIGHT, 1.f);
-
-        app::InputSystem::createBinding<player_move>(app::inputmap::method::W, 1.f);
-        app::InputSystem::createBinding<player_move>(app::inputmap::method::S, -1.f);
-        app::InputSystem::createBinding<player_strive>(app::inputmap::method::D, 1.f);
-        app::InputSystem::createBinding<player_strive>(app::inputmap::method::A, -1.f);
-        app::InputSystem::createBinding<player_fly>(app::inputmap::method::SPACE, 1.f);
-        app::InputSystem::createBinding<player_fly>(app::inputmap::method::LEFT_SHIFT, -1.f);
-        app::InputSystem::createBinding<player_look_x>(app::inputmap::method::MOUSE_X, 0.f);
-        app::InputSystem::createBinding<player_look_y>(app::inputmap::method::MOUSE_Y, 0.f);
-
-        app::InputSystem::createBinding<player_hover>(app::inputmap::method::Q, 1.f);
-        app::InputSystem::createBinding<player_hover>(app::inputmap::method::E, -1.f);
-
-        app::InputSystem::createBinding<exit_action>(app::inputmap::method::ESCAPE);
-        app::InputSystem::createBinding<fullscreen_action>(app::inputmap::method::F11);
-        app::InputSystem::createBinding<escape_cursor_action>(app::inputmap::method::TAB);
-        app::InputSystem::createBinding<vsync_action>(app::inputmap::method::F1);
 
         app::InputSystem::createBinding<audio_move>(app::inputmap::method::I, 1.f);
         app::InputSystem::createBinding<audio_move>(app::inputmap::method::K, -1.f);
@@ -303,26 +128,14 @@ public:
         app::InputSystem::createBinding<rewind_audio_source>(app::inputmap::method::BACKSPACE);
         app::InputSystem::createBinding<audio_test_input>(app::inputmap::method::SLASH);
 
-        app::InputSystem::createBinding< activate_CRtest0>(app::inputmap::method::KP_0);
-        app::InputSystem::createBinding< activate_CRtest1>(app::inputmap::method::KP_1);
         app::InputSystem::createBinding< activate_CRtest2>(app::inputmap::method::KP_2);
         app::InputSystem::createBinding< activate_CRtest3>(app::inputmap::method::KP_3);
 
         app::InputSystem::createBinding< activateFrictionTest >(app::inputmap::method::KP_4);
 
-        
+
         app::InputSystem::createBinding< extendedPhysicsContinue>(app::inputmap::method::M);
         app::InputSystem::createBinding<nextPhysicsTimeStepContinue>(app::inputmap::method::N);
-
-        bindToEvent<player_move, &TestSystem::onPlayerMove>();
-        bindToEvent<player_strive, &TestSystem::onPlayerStrive>();
-        bindToEvent<player_fly, &TestSystem::onPlayerFly>();
-        bindToEvent<player_look_x, &TestSystem::onPlayerLookX>();
-        bindToEvent<player_look_y, &TestSystem::onPlayerLookY>();
-        bindToEvent<exit_action, &TestSystem::onExit>();
-        bindToEvent<fullscreen_action, &TestSystem::onFullscreen>();
-        bindToEvent<escape_cursor_action, &TestSystem::onEscapeCursor>();
-        bindToEvent<vsync_action, &TestSystem::onVSYNCSwap>();
 
         bindToEvent<physics_test_move, &TestSystem::onUnitPhysicsUnitTestMove>();
 
@@ -336,13 +149,11 @@ public:
         bindToEvent<stop_audio_source, &TestSystem::stopAudioSource>();
         bindToEvent<rewind_audio_source, &TestSystem::rewindAudioSource>();
 
-        bindToEvent<physics::TriggerEvent,&TestSystem::testPhysicsEvent>();
+        bindToEvent<physics::TriggerEvent, &TestSystem::testPhysicsEvent>();
 
         bindToEvent<audio_test_input, &TestSystem::audioTestInput>();
 
         //collision resolution test
-        bindToEvent< activate_CRtest0, &TestSystem::onActivateUnitTest0>();
-        bindToEvent< activate_CRtest1, &TestSystem::onActivateUnitTest1>();
         bindToEvent< activate_CRtest2, &TestSystem::onActivateUnitTest2>();
         bindToEvent< activate_CRtest3, &TestSystem::onActivateUnitTest3>();
 
@@ -352,14 +163,8 @@ public:
 
         bindToEvent< extendedPhysicsContinue, &TestSystem::onExtendedPhysicsContinueRequest>();
         bindToEvent<nextPhysicsTimeStepContinue, &TestSystem::onNextPhysicsTimeStepRequest>();
-        bindToEvent<player_hover, &TestSystem::onPlayerHover>();
 
-        
 #pragma endregion
-
-        app::window window = m_ecs->world.get_component_handle<app::window>().read();
-        window.enableCursor(false);
-        window.show();
 
 #pragma region Model and material loading
         rendering::model_handle directionalLightH;
@@ -397,6 +202,8 @@ public:
         rendering::material_handle skyboxH;
         rendering::material_handle floorMH;
 
+        app::window window = m_ecs->world.get_component_handle<app::window>().read();
+
         {
             async::readwrite_guard guard(*window.lock);
             app::ContextHelper::makeContextCurrent(window);
@@ -421,10 +228,10 @@ public:
             auto colorshader = rendering::ShaderCache::create_shader("color", "assets://shaders/color.shs"_view);
             directionalLightMH = rendering::MaterialCache::create_material("directional light", colorshader);
             directionalLightMH.set_param("color", math::color(1, 1, 0.8f));
-            
+
             spotLightMH = rendering::MaterialCache::create_material("spot light", colorshader);
             spotLightMH.set_param("color", math::colors::green);
-            
+
             pointLightMH = rendering::MaterialCache::create_material("point light", colorshader);
             pointLightMH.set_param("color", math::colors::red);
 
@@ -592,13 +399,13 @@ public:
         {
             auto ent = createEntity();
             ent.add_components<rendering::renderable, rendering::light>({ directionalLightH, directionalLightMH }, rendering::light::directional(math::color(1, 1, 0.8f), 10.f));
-            ent.add_components<transform>(position(5, 13, 5), rotation::lookat(math::vec3(1,1,1), math::vec3::zero), scale());
+            ent.add_components<transform>(position(5, 13, 5), rotation::lookat(math::vec3(1, 1, 1), math::vec3::zero), scale());
         }
 
         {
             auto ent = createEntity();
             ent.add_components<rendering::renderable, rendering::light>({ spotLightH, spotLightMH }, rendering::light::spot(math::colors::green, math::deg2rad(90.f), 10.f));
-            ent.add_components<transform>(position(0.1f, 10, -7), rotation::lookat(math::vec3(0,0,-1), math::vec3::zero), scale());
+            ent.add_components<transform>(position(0.1f, 10, -7), rotation::lookat(math::vec3(0, 0, -1), math::vec3::zero), scale());
         }
 
         {
@@ -684,15 +491,15 @@ public:
             ent.add_components<transform>(position(4, 3, -5.1f), rotation(), scale(2.5f));
         }
 
-        math::mat4 rotMatrix = math::rotate(1.5f * 3.14159265f, math::vec3(0,1,0));
+        math::mat4 rotMatrix = math::rotate(1.5f * 3.14159265f, math::vec3(0, 1, 0));
         //audioSphereLeft setup
         {
-            audioSphereLeft = createEntity(); 
+            audioSphereLeft = createEntity();
             audioSphereLeft.add_component<rendering::renderable>({ cubeH, normalH });
             audioSphereLeft.add_components<transform>(position(-5, 0, 10), rotation(rotMatrix), scale(0.5));
 
             auto segment = audio::AudioSegmentCache::createAudioSegment("kilogram", "assets://audio/kilogram-of-scotland_stereo32.wav"_view, { audio::audio_import_settings::channel_processing_setting::split_channels });
-            
+
             audio::audio_source source;
             source.setAudioHandle(segment);
             audioSphereLeft.add_component<audio::audio_source>(source);
@@ -711,7 +518,7 @@ public:
         }
 #pragma endregion
 
-        
+
 
         //---------------------------------------------------------- Physics Collision Unit Test -------------------------------------------------------------------//
 
@@ -728,12 +535,12 @@ public:
         cubeParams.breadth = 1.0f;
         cubeParams.width = 1.0f;
         cubeParams.height = 1.0f;
-       //setupPhysicsCRUnitTest(cubeH, uvH);
+        //setupPhysicsCRUnitTest(cubeH, uvH);
 
 
-     auto sceneEntity = createEntity();
+        auto sceneEntity = createEntity();
         std::vector<ecs::entity_handle> children;
-        for (size_type i = 0;i < m_ecs->world.child_count();i++)
+        for (size_type i = 0; i < m_ecs->world.child_count(); i++)
         {
             children.push_back(m_ecs->world.get_child(i));
         }
@@ -746,26 +553,21 @@ public:
         }
 
         scenemanagement::SceneManager::createScene("Main", sceneEntity);
-  
+
         //sceneEntity.destroy();
 
         //scenemanagement::SceneManager::loadScene("ImposterFlake");
-
-        setupCameraEntity();
-        
 
         //CreateCubeStack(3, 2, 2, math::vec3(0, -3.0f, 8.0f), math::vec3(1, 1, 1)
         //    ,cubeParams, 0.1f, cubeH, wireframeH);
 
         createProcess<&TestSystem::update>("Update");
-        createProcess<&TestSystem::differentThread>("TestChain");
-        createProcess<&TestSystem::differentInterval>("TestChain", 1.f);
         //createProcess<&TestSystem::drawInterval>("TestChain");
     }
 
     void testPhysicsEvent(physics::TriggerEvent* evnt)
     {
-        log::debug("received trigger event {}",evnt->manifold.isColliding);
+        log::debug("received trigger event {}", evnt->manifold.isColliding);
     }
 
     void setupPhysicsCDUnitTest(rendering::model_handle cubeH, rendering::material_handle wireframeH)
@@ -1063,7 +865,7 @@ public:
 
             auto [positionH, rotationH, scaleH] = m_ecs->createComponents<transform>(ent);
             positionH.write(math::vec3(testPos, -3.0f, 15.0f));
-            scaleH.write(math::vec3(2.5f,1.0f,2.5f));
+            scaleH.write(math::vec3(2.5f, 1.0f, 2.5f));
 
             auto renderableHandle = m_ecs->createComponent<rendering::renderable>(ent);
             renderableHandle.write({ cubeH, wireframeH });
@@ -1212,7 +1014,7 @@ public:
             //renderableHandle.write({ cubeH, wireframeH });
 
             auto [positionH, rotationH, scaleH] = m_ecs->createComponents<transform>(staticToAABBEntRotation);
-            positionH.write(math::vec3(testPos-2.7f, -0.0f, 8.0f));
+            positionH.write(math::vec3(testPos - 2.7f, -0.0f, 8.0f));
             scaleH.write(math::vec3(1.0f));
 
             auto idHandle = m_ecs->createComponent<physics::identifier>(staticToAABBEntRotation);
@@ -1242,7 +1044,7 @@ public:
             physics::physicsComponent physicsComponent2;
             physics::physicsComponent::init(physicsComponent2);
 
-            
+
 
             physicsComponent2.AddBox(staticBlockParams);
             physicsComponent2.isTrigger = false;
@@ -1294,7 +1096,7 @@ public:
             id.id = "OBBRb";
             idHandle.write(id);
 
-            
+
             //log::debug("rb.angularVelocity {}", rb.angularVelocity);
 
         }
@@ -1446,7 +1248,7 @@ public:
             //renderableHandle.write({ cubeH, wireframeH });
 
             auto [positionH, rotationH, scaleH] = m_ecs->createComponents<transform>(NoFrictionBody);
-            positionH.write(math::vec3(testPos-1.0f, -1.5f, 15.0f));
+            positionH.write(math::vec3(testPos - 1.0f, -1.5f, 15.0f));
             scaleH.write(math::vec3(1.0f));
 
             auto idHandle = m_ecs->createComponent<physics::identifier>(NoFrictionBody);
@@ -1693,7 +1495,7 @@ public:
 
         {
             auto ent = m_ecs->createEntity();
-         
+
             auto entPhyHande = ent.add_component<physics::physicsComponent>();
 
             physics::physicsComponent physicsComponent2;
@@ -1720,7 +1522,7 @@ public:
         }
 
         CreateCubeStack(2, 1, 1, math::vec3(35, -1.5f, 15.0f), math::vec3(1, 1, 1)
-                ,cubeParams, 0.1f, cubeH, wireframeH);
+            , cubeParams, 0.1f, cubeH, wireframeH);
 
         //10 stactk
 
@@ -1728,80 +1530,6 @@ public:
     }
 
 #pragma region input stuff
-    void setupCameraEntity()
-    {
-        player = createEntity();
-        player.add_components<transform>(position(0.f, 3.f, 0.f), rotation::lookat(math::vec3::zero, math::vec3::forward), scale());
-        player.add_component<audio::audio_listener>();
-
-        rendering::camera cam;
-        cam.set_projection(90.f, 0.1f, 1000.f);
-        player.add_component<rendering::camera>(cam);
-    }
-
-    void onExit(exit_action* action)
-    {
-        if (action->released())
-            raiseEvent<events::exit>();
-    }
-
-    void onFullscreen(fullscreen_action* action)
-    {
-        if (action->released())
-        {
-            app::WindowSystem::requestFullscreenToggle(world_entity_id, math::ivec2(100, 100), math::ivec2(1360, 768));
-        }
-    }
-
-    void onEscapeCursor(escape_cursor_action* action)
-    {
-        if (action->released())
-        {
-            static bool enabled = false;
-            app::window window = m_ecs->world.get_component_handle<app::window>().read();
-            enabled = !enabled;
-            window.enableCursor(enabled);
-            window.show();
-        }
-    }
-
-    void onVSYNCSwap(vsync_action* action)
-    {
-        if (action->released())
-        {
-            auto handle = m_ecs->world.get_component_handle<app::window>();
-            app::window window = handle.read();
-            window.setSwapInterval(window.swapInterval() ? 0 : 1);
-            log::debug("set swap interval to {}", window.swapInterval());
-            handle.write(window);
-        }
-    }
-
-    void onPlayerMove(player_move* action)
-    {
-        auto posH = player.get_component_handle<position>();
-        auto rot = player.get_component_handle<rotation>().read();
-        math::vec3 move = math::toMat3(rot) * math::vec3(0.f, 0.f, 1.f);
-        move = math::normalize(move * math::vec3(1, 0, 1)) * action->value * action->input_delta * 10.f;
-        posH.fetch_add(move);
-    }
-
-    void onPlayerHover(player_hover* action)
-    {
-        auto posH = player.get_component_handle<position>();
-        math::vec3 move = math::vec3(0, 1, 0) * action->value * 0.01f;
-        posH.fetch_add(move);
-    }
-
-    void onPlayerStrive(player_strive* action)
-    {
-        auto posH = player.get_component_handle<position>();
-        auto rot = player.get_component_handle<rotation>().read();
-        math::vec3 move = math::toMat3(rot) * math::vec3(1.f, 0.f, 0.f);
-        move = math::normalize(move * math::vec3(1, 0, 1)) * action->value * action->input_delta * 10.f;
-        posH.fetch_add(move);
-    }
-
     void onSphereAAMove(audio_move* action)
     {
         auto posH = audioSphereLeft.get_component_handle<position>();
@@ -1847,48 +1575,6 @@ public:
         audio_source source = sourceH.read();
         source.disableSpatialAudio();
         sourceH.write(source);
-    }
-
-    void onPlayerFly(player_fly* action)
-    {
-        auto posH = player.get_component_handle<position>();
-        posH.fetch_add(math::vec3(0.f, action->value * action->input_delta * 10.f, 0.f));
-    }
-
-    void onPlayerLookX(player_look_x* action)
-    {
-        auto rotH = player.get_component_handle<rotation>();
-        rotH.fetch_multiply(math::angleAxis(action->value, math::vec3(0, 1, 0)));
-        rotH.read_modify_write(rotation(), [](const rotation& src, rotation&& dummy)
-            {
-                (void)dummy;
-                math::vec3 fwd = math::toMat3(src) * math::vec3(0.f, 0.f, 1.f);
-                if (fwd.y < -0.95f)
-                    fwd.y = -0.95f;
-                else if (fwd.y > 0.95f)
-                    fwd.y = 0.95f;
-                fwd = math::normalize(fwd);
-                math::vec3 right = math::cross(fwd, math::vec3(0.f, 1.f, 0.f));
-                return (rotation)math::conjugate(math::toQuat(math::lookAt(math::vec3(0.f, 0.f, 0.f), fwd, math::cross(right, fwd))));
-            });
-    }
-
-    void onPlayerLookY(player_look_y* action)
-    {
-        auto rotH = player.get_component_handle<rotation>();
-        rotH.fetch_multiply(math::angleAxis(action->value, math::vec3(1, 0, 0)));
-        rotH.read_modify_write(rotation(), [](const rotation& src, rotation&& dummy)
-            {
-                (void)dummy;
-                math::vec3 fwd = math::toMat3(src) * math::vec3(0.f, 0.f, 1.f);
-                if (fwd.y < -0.95f)
-                    fwd.y = -0.95f;
-                else if (fwd.y > 0.95f)
-                    fwd.y = 0.95f;
-                fwd = math::normalize(fwd);
-                math::vec3 right = math::cross(fwd, math::vec3(0.f, 1.f, 0.f));
-                return (rotation)math::conjugate(math::toQuat(math::lookAt(math::vec3(0.f, 0.f, 0.f), fwd, math::cross(right, fwd))));
-            });
     }
 
     void onUnitPhysicsUnitTestMove(physics_test_move* action)
@@ -1969,31 +1655,17 @@ public:
         debug::drawLine(math::vec3(0, 0, 0), math::vec3(0, 1, 0), math::colors::green, 10);
         debug::drawLine(math::vec3(0, 0, 0), math::vec3(0, 0, 1), math::colors::blue, 10);
 
-        //log::info("still alive! {}",deltaTime.seconds());
         static auto sahQuery = createQuery<sah, rotation, position>();
-
-        //static time::span buffer;
-        static int frameCount;
-        //static time::span accumulated;
-
-        //buffer += deltaTime;
-        //accumulated += deltaTime;
-        frameCount++;
 
         for (auto entity : sahQuery)
         {
-            auto comp = entity.get_component_handle<sah>();
-
             auto rot = entity.read_component<rotation>();
 
             rot *= math::angleAxis(math::deg2rad(45.f * deltaTime), math::vec3(0, 1, 0));
 
             entity.write_component(rot);
 
-            comp.write({ frameCount });
-
             auto pos = entity.read_component<position>();
-
             debug::drawLine(pos, pos + rot.forward(), math::colors::magenta, 10);
         }
 
@@ -2008,73 +1680,6 @@ public:
                 entity.write_component(rot);
             }
         }
-
-        //if (buffer > 1.f)
-        //{
-        //    buffer -= 1.f;
-
-        //    for (auto entity : query)
-        //    {
-        //        auto comp = entity.get_component_handle<sah>();
-        //        std::cout << "component value: " << comp.read().value << std::endl;
-        //    }
-
-        //    std::cout << "Hi! " << (frameCount / accumulated) << "fps " << deltaTime.milliseconds() << "ms" << std::endl;
-        //}
-    }
-
-    void differentInterval(time::span deltaTime)
-    {
-        static time::span buffer;
-        static int frameCount;
-        static time::span accumulated;
-
-        buffer += deltaTime;
-        accumulated += deltaTime;
-        frameCount++;
-
-        math::vec2 v;
-        v.x = 10;
-        v.y = 20;
-
-        if (buffer > 1.f)
-        {
-            buffer -= 1.f;
-            //std::cout << "This is a fixed interval!! " << (frameCount / accumulated) << "fps " << deltaTime.milliseconds() << "ms" << std::endl;
-        }
-    }
-
-    void differentThread(time::span deltaTime)
-    {
-        static auto query = createQuery<sah>();
-
-        static time::span buffer;
-        static int frameCount;
-        static time::span accumulated;
-
-        buffer += deltaTime;
-        accumulated += deltaTime;
-        frameCount++;
-
-        /* if (buffer > 1.f)
-         {
-             buffer -= 1.f;
-
-             for (auto entity : query)
-             {
-                 auto comp = entity.get_component_handle<sah>();
-                 std::cout << "component value on different thread: " << comp.read().value << std::endl;
-             }
-
-             std::cout << "This is a different thread!! " << (frameCount / accumulated) << "fps " << deltaTime.milliseconds() << "ms" << std::endl;
-         }*/
-
-         //if (accumulated > 10.f)
-         //{
-         //	std::cout << "raising exit event" << std::endl;
-         //	raiseEvent<events::exit>();
-         //	//throw legion_exception_msg("hehehe fuck you >:D");
-         //}
     }
 
     void drawInterval(time::span deltaTime)
@@ -2084,58 +1689,55 @@ public:
 
         float duration = 0.02f;
 
-        //for (auto penetration : physics::PhysicsSystem::penetrationQueries)
-        //{
-        //    debug::drawLine(penetration->faceCentroid
-        //        , penetration->faceCentroid + penetration->normal, math::vec4(1, 0, 1, 1), 15.0f, duration);
-        //    auto x = 1;
-        //}
+        for (auto penetration : physics::PhysicsSystem::penetrationQueries)
+        {
+            debug::drawLine(penetration->faceCentroid
+                , penetration->faceCentroid + penetration->normal, math::vec4(1, 0, 1, 1), 15.0f, duration);
+            auto x = 1;
+        }
 
 
         //--------------------------------------- Draw contact points ---------------------------------------//
 
-       
 
-        for(int i = 0; i < physics::PhysicsSystem::contactPoints.size();i++)
+
+        for (int i = 0; i < physics::PhysicsSystem::contactPoints.size(); i++)
         {
             //ref is red
             //inc is blue
 
             auto& contact = physics::PhysicsSystem::contactPoints.at(i);
 
-   /*         debug::drawLine(contact.IncWorldContact
-                , contact.IncWorldContact + contact.totaldebugAddedAngular, math::vec4(1, 0, 1, 1), 5.0f, duration);*/
-
             debug::drawLine(contact.refRBCentroid
-                , contact.RefWorldContact, math::vec4(1, 0, 0, 1), 5.0f, duration,true);
+                , contact.RefWorldContact, math::vec4(1, 0, 0, 1), 5.0f, duration, true);
 
             debug::drawLine(contact.incRBCentroid
-                , contact.IncWorldContact, math::vec4(0, 0, 1, 1), 5.0f, duration,true);
+                , contact.IncWorldContact, math::vec4(0, 0, 1, 1), 5.0f, duration, true);
 
-      /*      debug::drawLine(contact.IncWorldContact
-                , contact.IncWorldContact + math::vec3(0, 0.1f, 0), math::vec4(0.5, 0.5, 0.5, 1),5.0f, duration, true);
+            debug::drawLine(contact.IncWorldContact
+                , contact.IncWorldContact + math::vec3(0, 0.1f, 0), math::vec4(0.5, 0.5, 0.5, 1), 5.0f, duration, true);
 
             debug::drawLine(contact.refRBCentroid
-                , contact.refRBCentroid + math::vec3(0, 0.1f, 0), math::vec4(0, 0, 0, 1), 5.0f, duration, true);*/
+                , contact.refRBCentroid + math::vec3(0, 0.1f, 0), math::vec4(0, 0, 0, 1), 5.0f, duration, true);
 
         }
 
         //--------------------------------------- Draw extreme points ---------------------------------------//
 
-        //i = 0;
-        //for (auto penetration : physics::PhysicsSystem::aPoint)
-        //{
-        //    debug::drawLine(penetration
-        //        , penetration + math::vec3(0, 0.2, 0), math::vec4(1, 0, 0, 1), 15.0f);
+        i = 0;
+        for (auto penetration : physics::PhysicsSystem::aPoint)
+        {
+            debug::drawLine(penetration
+                , penetration + math::vec3(0, 0.2, 0), math::vec4(1, 0, 0, 1), 15.0f);
 
-        //}
-        //i = 0;
-        //for (auto penetration : physics::PhysicsSystem::bPoint)
-        //{
-        //    debug::drawLine(penetration
-        //        , penetration + math::vec3(0, 0.2, 0), math::vec4(0, 0, 1, 1), 15.0f);
+        }
+        i = 0;
+        for (auto penetration : physics::PhysicsSystem::bPoint)
+        {
+            debug::drawLine(penetration
+                , penetration + math::vec3(0, 0.2, 0), math::vec4(0, 0, 1, 1), 15.0f);
 
-        //}
+        }
 
         physics::PhysicsSystem::contactPoints.clear();
         physics::PhysicsSystem::penetrationQueries.clear();
@@ -2192,27 +1794,22 @@ public:
                         math::vec3 faceStart = localTransform * math::vec4(face->centroid, 1);
                         math::vec3 faceEnd = faceStart + math::vec3((localTransform * math::vec4(face->normal, 0)));
 
-                        //debug::drawLine(faceStart, faceEnd, math::colors::green, 5.0f);
+                        debug::drawLine(faceStart, faceEnd, math::colors::green, 5.0f);
 
                         if (!currentEdge) { return; }
 
-                        //log::debug(("---- face"));
                         do
                         {
-                            //log::debug(("edge"));
                             physics::HalfEdgeEdge* edgeToExecuteOn = currentEdge;
                             currentEdge = currentEdge->nextEdge;
 
                             math::vec3 worldStart = localTransform * math::vec4(*(edgeToExecuteOn->edgePositionPtr), 1);
                             math::vec3 worldEnd = localTransform * math::vec4(*(edgeToExecuteOn->nextEdge->edgePositionPtr), 1);
 
-                            debug::drawLine(worldStart, worldEnd, usedColor, 2.0f,0.0f, useDepth);
+                            debug::drawLine(worldStart, worldEnd, usedColor, 2.0f, 0.0f, useDepth);
 
                         } while (initialEdge != currentEdge && currentEdge != nullptr);
                     }
-
-
-                    physCollider->collisionsFound.clear();
                 }
 
             }
@@ -2222,7 +1819,7 @@ public:
         //FindClosestPointsToLineSegment unit test
 
 
-       /* math::vec3 p1(5, -0.5, 0);
+        math::vec3 p1(5, -0.5, 0);
         math::vec3 p2(5, 0.5, 0);
 
         math::vec3 p3(6, 0, -0.5);
@@ -2249,54 +1846,7 @@ public:
 
         physics::PhysicsStatics::FindClosestPointsToLineSegment(p1, p2, p3, p4, p1p2, p3p4);
 
-        debug::drawLine(p1p2, p3p4, math::colors::green, 5.0f);*/
-
-        //log::debug("{:.3f}", deltaTime);
-    
-       
-    }
-
-    void onActivateUnitTest1(activate_CRtest1* action)
-    {
-        static bool activated = false;
-
-        //auto [qpositionH, qrotationH, qscaleH] = m_ecs->getComponents<transform>(staticToAABBEnt);
-        //math::vec3 qpos = qpositionH.read();
-        //log::debug("BEFORE ADD rbPos {} ", math::to_string(qpos));
-
-        if (action->value && !activated)
-        {
-            activated = true;
-            auto crb = m_ecs->createComponent<physics::rigidbody>(staticToAABBEntRotation);
-            //auto rb = crb.read();
-
-            //auto [positionH, rotationH, scaleH] = m_ecs->getComponents<transform>(staticToAABBEnt);
-            //math::vec3 pos = positionH.read();
-            //log::debug("ON ADD rbPos {} ", math::to_string(pos));
-        }
-
-     
-    }
-
-    void onActivateUnitTest0(activate_CRtest0* action)
-    {
-        static bool activated = false;
-
-        //auto [qpositionH, qrotationH, qscaleH] = m_ecs->getComponents<transform>(staticToAABBEnt);
-        //math::vec3 qpos = qpositionH.read();
-        //log::debug("BEFORE ADD rbPos {} ", math::to_string(qpos));
-
-        if (action->value && !activated)
-        {
-            activated = true;
-            auto crb = m_ecs->createComponent<physics::rigidbody>(staticToAABBEntLinear);
-            //auto rb = crb.read();
-
-            //auto [positionH, rotationH, scaleH] = m_ecs->getComponents<transform>(staticToAABBEnt);
-            //math::vec3 pos = positionH.read();
-            //log::debug("ON ADD rbPos {} ", math::to_string(pos));
-        }
-
+        debug::drawLine(p1p2, p3p4, math::colors::green, 5.0f);
 
     }
 
@@ -2319,7 +1869,7 @@ public:
         }
     }
 
-    void onExtendedPhysicsContinueRequest(extendedPhysicsContinue * action)
+    void onExtendedPhysicsContinueRequest(extendedPhysicsContinue* action)
     {
         if (action->value)
         {
@@ -2344,7 +1894,7 @@ public:
     }
 
 
-    void FrictionTestActivate(activateFrictionTest * action)
+    void FrictionTestActivate(activateFrictionTest* action)
     {
         if (action->value)
         {
@@ -2358,7 +1908,7 @@ public:
 
                 rbH.write(rb);
             }
-            
+
             {
                 auto rbH = m_ecs->createComponent<physics::rigidbody>(Point3FrictionBody);
 
@@ -2414,7 +1964,7 @@ public:
                     entPhyHande.write(physicsComponent);
 
                     auto [positionH, rotationH, scaleH] = m_ecs->createComponents<transform>(ent);
-                    positionH.write(startPosition + math::vec3(j * offset.x , i * offset.y , k * offset.z));
+                    positionH.write(startPosition + math::vec3(j * offset.x, i * offset.y, k * offset.z));
                     scaleH.write(math::vec3(1.0f));
 
                     if (addRigidbody)
